feat(rich-text-editor): refactor rich text editor components and update serialization logic
